WEBREF-116 :truck: Moved global style to own module
