+++ src/layout/index.ts
@@ -1,2 +1,1 @@
 export { default as Layout } from './Layout'
-export { GlobalStyle } from './Layout.styles'
